Created Woog Domain
This is the model compiler domain. It's work towards #4.

I had to hack on nut to make this work, and even then it's only sort
of working. It uses a relationship type I'd never generated macros,
not to mention macro tests, for. That reflexive relationship also
wreaked havoc with sarzak. Stack overflow. Lot's of fun.