re-org
not much to say that one cannot see for themselves.

the impetus was that i, for fun, changed sarzak to export woog v2. I
wanted to see how involved it would be to switch over to it. domain
broke. I realized I'd need more, where to put it? in v1? Why isn't
everything else in there?

Lesson? Don't put underscores in directory names. Make
subdirectories.