Update Sarzak Model
R15 is now many on the subtype side, since an object may be a subtype
to many supertypes. I honestly dismissed this as rational. Until I
needed it. It's busted BTW. I can't generate code with it using the
current model and type system. I should be able to pull it off with
woog though. I think/hope!