Dos cosas
I updated the merlin model with some object
descriptions. That poor thing is the result of
some manic session. It's pretty brilliant, as I
recall, but I didn't document it, and I'm not
sure how it works. Needs work. 🚧

I also changed all the compiler bits to pass
around references to Arc's. That got rid of a
bunch of clonening, so yay! 🎉