I don't know what I'm doing anymore.
This has a lot of changes. I think that I wanted to slim down the repo.
There is also a change to lu_dog making r4 associative. I also switched
the hashmap implementation to `rustc_hash`. There is also a change to
`sarzak.toml` adding a "no deadlocks" version of rwlock to dwarf.