Store Update
Updated everything to be using at least a single uber-store. So that's
`Rc<RefCell<Z>>`, where `Z` is an instance from sarzak. This is way
better than what I had originally done. It's a big change though. It
was a lot to fix dwarf, and I think that grace is going to be neigh
unbearable. If I choose to go that route.

Why did I do this? I'm not really sure. I think mostly because the plug-
in work I'm doing is based on sarzak. And it felt weird not having things
at least Rc. Is that dumb? Is this really better? Why am I questioning
myself now?

Sigh -- it's always possible to go back. Wait, wait. Hold on. This is
dwarf. We need this stuff to be Rc so that we can twiddle references
without having to worry. We get a lot for free by having interior
mutability and reference counting.

Grace doesn't really need any of that, and while sometimes it's gross,
it's fast as shit on a bad morning. 🚽💩 Someone please make him stop.